# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Kassomat API
  version: 1.0.0
  description: |
    Kassomat POS SaaS API — RKSV-konforme Registrierkasse für österreichische Betriebe.

    ## Authentifizierung
    Alle Endpoints (außer /auth/* und /webhooks/*) erfordern einen Bearer Token im
    `Authorization` Header: `Authorization: Bearer <access_token>`

    ## Multi-Tenant
    Alle Daten sind tenant-isoliert. Der Tenant wird aus dem JWT extrahiert.
    Ein Benutzer kann niemals Daten eines anderen Tenants sehen oder verändern.

    ## Geldbeträge
    Alle Geldbeträge sind in **Cent (Integer)**. Beispiel: €2,50 = `250`
servers:
  - url: https://api.kassomat.at
    description: Production
  - url: https://api-staging.kassomat.at
    description: Staging
  - url: http://localhost:3001
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ============================================================
  # REUSABLE SCHEMAS
  # ============================================================
...
security:
  - bearerAuth: []

# ============================================================
# PATHS
# ============================================================
paths:
...
  /webhooks/lieferando:
    post:
      summary: Eingehender Webhook von Lieferando / JET POS API
      security: []
      tags: [Webhooks]
      parameters:
        - name: X-JET-Signature
          in: header
          required: true
          schema:
            type: string
          description: HMAC-SHA256 Signatur des Payloads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JET POS API Webhook Payload
      responses:
        '200':
          description: Webhook verarbeitet
        '401':
          description: Ungültige Signatur
  /webhooks/wix:
    post:
      summary: Eingehender Webhook von Wix
      security: []
      tags: [Webhooks]
      parameters:
        - name: X-Wix-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Wix eCommerce Order Webhook Payload
      responses:
        '200':
          description: Webhook verarbeitet
        '401':
          description: Ungültige Signatur

  # ---- SHIFTS ----
...